backend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
schedule==1.2.0
selenium==4.15.2
webdriver-manager==4.0.1
